Rename context from cam to query
