--- src/context/camContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const camContext = createContext();
-
-export default camContext;
