+++ src/context/queryContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const queryContext = createContext();
+
+export default queryContext;
